Update setup.py
This I think will stop the wrong pyart package from being installed(unless it ever crosses version 1.0). 

This PR changes there require to request a version of pyart that is greater than the python adaptive radix tree currently on PyPI. This should make an install without the atmos pyart library fail, and stop pip from thinking it can fulfill the requirement with the wrong package. 
If the atmos pyart package is installed, I believe this should not change anything.